+++ Math/src/Math_py.tpp
@@ -183,7 +183,7 @@ template<class DataType>
 DataType sum_pairwise_py(py::array_t<DataType,py::array::c_style>& py_in1)
 {
 	py::buffer_info buf1 = py_in1.request();
-	int n = buf1.size;
+	long int n = buf1.size;
 	DataType* in = (DataType*) buf1.ptr;
 	return sum_pairwise(in,n);
 }
